requirements.txt
Django==4.1.1
django_multiselectfield==0.1.12
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.0
